# internal/application/modules/categories/dtos/category_dto.go
package dtos

import "time"

type CategoryDto struct {
	ID          int64     `json:"id"`
	Name        string    `json:"name"`
	Description string    `json:"description"`
	CreatedAt   time.Time `json:"createdAt"`
}